added contact pag
