+++ src/pages/pf-contact/index.jsx
@@ -1,0 +1,3 @@
+import { Contact } from "./contact.container";
+
+export default Contact;
